refactor: rework game layout creation
